moneymong-323 refactor: 불필요한 코드 삭제 및 패키지 네이밍 변경
