+++ data/src/main/java/com/moneymong/moneymong/data/di/DataSourceModule.kt
@@ -2,14 +2,6 @@ package com.moneymong.moneymong.data.di
 
 import com.moneymong.moneymong.data.datasource.agency.AgencyLocalDataSource
 import com.moneymong.moneymong.data.datasource.agency.AgencyLocalDataSourceImpl
-import com.moneymong.moneymong.data.datasource.login.LoginLocalDataSource
-import com.moneymong.moneymong.data.datasource.login.LoginLocalDataSourceImpl
-import com.moneymong.moneymong.data.datasource.login.LoginRemoteDataSource
-import com.moneymong.moneymong.data.datasource.login.LoginRemoteDataSourceImpl
-import com.moneymong.moneymong.data.datasource.login.TokenRemoteDataSource
-import com.moneymong.moneymong.data.datasource.login.TokenRemoteDataSourceImpl
-import com.moneymong.moneymong.data.datasource.signup.UnivRemoteDataSource
-import com.moneymong.moneymong.data.datasource.signup.UnivRemoteDataSourceImpl
 import com.moneymong.moneymong.data.datasource.agency.AgencyRemoteDataSource
 import com.moneymong.moneymong.data.datasource.agency.AgencyRemoteDataSourceImpl
 import com.moneymong.moneymong.data.datasource.ledger.LedgerLocalDataSource
@@ -18,10 +10,16 @@ import com.moneymong.moneymong.data.datasource.ledger.LedgerRemoteDataSource
 import com.moneymong.moneymong.data.datasource.ledger.LedgerRemoteDataSourceImpl
 import com.moneymong.moneymong.data.datasource.ledgerdetail.LedgerDetailRemoteDataSource
 import com.moneymong.moneymong.data.datasource.ledgerdetail.LedgerDetailRemoteDataSourceImpl
+import com.moneymong.moneymong.data.datasource.login.LoginLocalDataSource
+import com.moneymong.moneymong.data.datasource.login.LoginLocalDataSourceImpl
+import com.moneymong.moneymong.data.datasource.login.TokenRemoteDataSource
+import com.moneymong.moneymong.data.datasource.login.TokenRemoteDataSourceImpl
 import com.moneymong.moneymong.data.datasource.member.MemberRemoteDataSource
 import com.moneymong.moneymong.data.datasource.member.MemberRemoteDataSourceImpl
 import com.moneymong.moneymong.data.datasource.ocr.OCRRemoteDataSource
 import com.moneymong.moneymong.data.datasource.ocr.OCRRemoteDataSourceImpl
+import com.moneymong.moneymong.data.datasource.signup.UnivRemoteDataSource
+import com.moneymong.moneymong.data.datasource.signup.UnivRemoteDataSourceImpl
 import com.moneymong.moneymong.data.datasource.user.UserLocalDataSource
 import com.moneymong.moneymong.data.datasource.user.UserLocalDataSourceImpl
 import com.moneymong.moneymong.data.datasource.user.UserRemoteDataSource
@@ -34,11 +32,6 @@ import dagger.hilt.components.SingletonComponent
 @Module
 @InstallIn(SingletonComponent::class)
 interface DataSourceModule {
-
-    @Binds
-    fun bindLoginDataSource(
-        loginRemoteDataSourceImpl: LoginRemoteDataSourceImpl
-    ): LoginRemoteDataSource
 
     @Binds
     fun bindUnivDataSource(
