--- domain/src/main/java/com/moneymong/moneymong/domain/repository/LoginRepository.kt
@@ -1,7 +1,0 @@
-package com.moneymong.moneymong.domain.repository
-
-import com.moneymong.moneymong.model.sign.LoginType
-
-interface LoginRepository {
-    suspend fun postAccessToken(type: LoginType, accessToken: String): Result<Unit>
-}
